+++ assets/js/changeHP.js
@@ -1,0 +1,9 @@
+function changeHP(randomDamage) {
+  this.hp -= randomDamage;
+
+  if (this.hp <= 0) {
+    this.hp = 0;
+  }
+}
+
+export default changeHP;
