+++ assets/js/createPlayer.js
@@ -1,0 +1,26 @@
+import createElement from "./createElement.js";
+
+function createPlayer(player) {
+  const playerBox = createElement('div', 'player' +player.player);
+  const progressBar  = createElement('div', 'progressbar');
+  const characterBox = createElement('div', 'character');
+  const life = createElement('div', 'life');
+  const name = createElement('div', 'name');
+  const img = createElement('img');
+
+  life.style.width = player.hp + '%';
+  img.src = player.img;
+  name.textContent = player.name;
+
+  playerBox.appendChild(progressBar);
+  playerBox.appendChild(characterBox);
+
+  progressBar.appendChild(life);
+  progressBar.appendChild(name);
+
+  characterBox.appendChild(img);
+
+  return playerBox;
+}
+
+export default createPlayer;
