+++ assets/js/elHP.js
@@ -1,0 +1,5 @@
+function elHP() {
+  return document.querySelector('.player' + this.player + ' .life');
+}
+
+export default elHP;
