+++ assets/js/getRandomFromRange.js
@@ -1,0 +1,8 @@
+function getRandomFromRange(min, max) {
+  min = Math.ceil(min);
+  max = Math.floor(max);
+
+  return Math.floor(Math.random() * (max - min + 1)) + min;
+}
+
+export default getRandomFromRange;
