+++ assets/js/getTime.js
@@ -1,0 +1,7 @@
+function getTime() {
+  const time = new Date;
+
+  return `${time.getHours()}:${time.getMinutes()}`
+}
+
+export default getTime;
