+++ assets/js/renderHP.js
@@ -1,0 +1,5 @@
+function renderHP() {
+  this.elHP().style.width = this.hp + '%';
+}
+
+export default renderHP;
